---  webpack.config.js
@@ -1,30 +1,0 @@
-module.exports = {
-  mode: 'development',
-  devtool: 'inline-source-map',
-  devServer: {
-    contentBase: './dist',
-  },
-  module: {
-    rules: [{
-      test: /\.s[ac]ss$/i,
-      use: [
-        // Creates `style` nodes from JS strings
-        'style-loader',
-        // Translates CSS into CommonJS
-        'css-loader',
-        // Compiles Sass to CSS
-        'sass-loader',
-      ],
-    },
-    {
-      test: /\.(png|svg|jpg|jpeg|gif)$/i,
-      type: 'assert/resource',
-    },
-    {
-      test: /\.html$/i,
-      loader: 'html-loader',
-    },
-
-    ],
-  },
-};
