+++ .eslintrc.json
@@ -16,8 +16,7 @@
     "no-shadow": "off",
     "no-param-reassign": "off",
     "eol-last": "off",
-    "arrow-parens": "off",
-    "no-alert":"off"
+    "arrow-parens": "off"
   },
   "ignorePatterns": [
     "dist/",
