+++ src/date.js
@@ -1,0 +1,15 @@
+
+const dateManage = (dateArg) => {
+  const days = ['Sunday', 'Monday', 'Thuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'];
+
+  const months = ['Junuary', 'February', 'March', 'April', 'May', 'Jun', 'July', 'August', 'September', 'October', 'November', 'December'];
+
+  const year = dateArg.getFullYear();
+  const month = months[dateArg.getMonth()];
+  const date = dateArg.getDate();
+  const day = days[dateArg.getDay()];
+
+  return `${date} ${month} (${day}), ${year}`;
+};
+
+export default dateManage;
