--- src/page/date.js
@@ -1,14 +1,0 @@
-const dateArray = (dateArg) => {
-  const days = ['Sunday', 'Monday', 'Thuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday'];
-
-  const months = ['Junuary', 'February', 'March', 'April', 'May', 'Jun', 'July', 'August', 'September', 'October', 'November', 'December'];
-
-  const year = dateArg.getFullYear();
-  const month = months[dateArg.getMonth()];
-  const date = dateArg.getDate();
-  const day = days[dateArg.getDay()];
-
-  return `${month} ${date} (${day}), ${year}`;
-};
-
-export default dateArray;
