--- src/page/temperature.js
@@ -1,3 +1,0 @@
-const tempToF = (temp) => temp * (9 / 5) + 32;
-
-export default tempToF;
