--- src/page/weather.js
@@ -1,40 +1,0 @@
-import dateArray from './date';
-import convertTemp from './toogle';
-
-const showWeatherReport = (weather) => {
-  const city = document.querySelector('#city');
-  city.innerText = `${weather.name}, ${weather.sys.country}`;
-
-  const temperature = document.querySelector('#temp');
-  temperature.innerHTML = `${Math.round(weather.main.temp)}&deg;C`;
-
-  const minMax = document.querySelector('#min-max');
-  minMax.innerHTML = `${Math.floor(weather.main.temp_min)}&deg;C (min) / ${Math.ceil(weather.main.temp_max)}&deg;C (max)`;
-
-  const weatherType = document.querySelector('#description');
-  weatherType.innerText = `${weather.weather[0].main}`;
-
-  const weatherIcon = document.querySelector('#icon');
-  weatherIcon.innerText = `${weather.weather[0].icon}`;
-
-  const date = document.querySelector('#date');
-  const newDate = new Date();
-  date.innerText = dateArray(newDate);
-  convertTemp(weather);
-
-  const weatherTypes = {
-    Clear: 'sun',
-    Clouds: 'cloudy',
-    Rain: 'rain',
-    Mist: 'fog',
-    Snow: 'snow',
-    Thunderstorms: 'thunder',
-    Haze: 'fog',
-  };
-
-  document.body.style.backgroundImage = `url('./images/${weatherTypes[weatherType.textContent]}.gif')`;
-
-  document.querySelector('#icon').src = `./icons/${weatherIcon.textContent}.png`;
-};
-
-export default showWeatherReport;
