+++ src/temperature.js
@@ -1,0 +1,3 @@
+const tempToF = (temp) => temp * (9 / 5) + 32;
+
+export default tempToF;
